413 error treated, entrega concluida

+++ app/__init__.py
@@ -1,6 +1,6 @@
 from http import HTTPStatus
 from flask import Flask, jsonify, request, send_file
-from app.kenzie import create_files_dir, retrieve_all_files, retrieve_extension_files, allowed_extensions_filter, upload_file, download_filepath
+from app.kenzie import create_files_dir, retrieve_all_files, retrieve_extension_files, allowed_extensions_filter, already_exists_filter, upload_file, download_filepath
 import os
 
 MAX_CONTENT_LENGTH = os.getenv("MAX_CONTENT_LENGTH")
@@ -19,8 +19,7 @@ def list_files():
 def list_files_by_extension(extension):
 
     if allowed_extensions_filter(extension):
-        return {'msg': f'extension {extension} not allowed or incorrect format'}, HTTPStatus.BAD_REQUEST
-
+        return {'msg': f'extension {extension} not allowed or incorrect format'}, HTTPStatus.NOT_FOUND
     return jsonify(retrieve_extension_files(extension)), HTTPStatus.OK
 
 @app.post("/upload")
@@ -28,7 +27,9 @@ def upload():
 
     for file in request.files.values():
         if allowed_extensions_filter(file.filename):
-            return {'msg': f"file's {file.filename} extension not allowed or incorrect format"}, HTTPStatus.BAD_REQUEST
+            return {'msg': f"file's {file.filename} extension not allowed or incorrect format"}, HTTPStatus.UNSUPPORTED_MEDIA_TYPE
+        if already_exists_filter(file.filename):
+            return {'msg': f"filename {file.filename} already exists"}, HTTPStatus.CONFLICT
 
         upload_file(file)
 
@@ -39,6 +40,10 @@ def download(file_name):
     try:
         return send_file(download_filepath(file_name), as_attachment=True), HTTPStatus.ACCEPTED
     except ValueError:
-        return {'msg': "invalid name"}, HTTPStatus.BAD_REQUEST
+        return {'msg': "invalid name"}, HTTPStatus.NOT_FOUND
     except FileNotFoundError:
-        return {'msg': "file does not exist"}, HTTPStatus.BAD_REQUEST
+        return {'msg': "file does not exist"}, HTTPStatus.NOT_FOUND
+
+@app.errorhandler(413)
+def REQUEST_ENTITY_TOO_LARGE(error):
+    return {'msg': 'file too large'}, HTTPStatus.REQUEST_ENTITY_TOO_LARGE
